Move dbal check
