+++ src/Illuminate/Foundation/Console/ShowModelCommand.php
@@ -7,6 +7,7 @@ use Doctrine\DBAL\Schema\Index;
 use Doctrine\DBAL\Types\DecimalType;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Container\BindingResolutionException;
+use Illuminate\Database\Console\AbstractDatabaseCommand;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\Composer;
 use Illuminate\Support\Str;
@@ -15,12 +16,9 @@ use ReflectionMethod;
 use SplFileObject;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Process\Exception\ProcessSignaledException;
-use Symfony\Component\Process\Exception\RuntimeException;
-use Symfony\Component\Process\Process;
 
 #[AsCommand(name: 'model:show')]
-class ShowModelCommand extends Command
+class ShowModelCommand extends AbstractDatabaseCommand
 {
     /**
      * The console command name.
@@ -102,13 +100,7 @@ class ShowModelCommand extends Command
      */
     public function handle()
     {
-        if (! interface_exists('Doctrine\DBAL\Driver')) {
-            if (! $this->components->confirm('Displaying model information requires the Doctrine DBAL (doctrine/dbal) package. Would you like to install it?')) {
-                return 1;
-            }
-
-            return $this->installDependencies();
-        }
+        $this->ensureDependenciesExist();
 
         $class = $this->qualifyModel($this->argument('model'));
 
@@ -173,9 +165,10 @@ class ShowModelCommand extends Command
         $class = new ReflectionClass($model);
 
         return collect($class->getMethods())
-            ->reject(fn (ReflectionMethod $method) => $method->isStatic()
-                || $method->isAbstract()
-                || $method->getDeclaringClass()->getName() !== get_class($model)
+            ->reject(
+                fn (ReflectionMethod $method) => $method->isStatic()
+                    || $method->isAbstract()
+                    || $method->getDeclaringClass()->getName() !== get_class($model)
             )
             ->mapWithKeys(function (ReflectionMethod $method) use ($model) {
                 if (preg_match('/^get(.*)Attribute$/', $method->getName(), $matches) === 1) {
@@ -212,9 +205,10 @@ class ShowModelCommand extends Command
     {
         return collect(get_class_methods($model))
             ->map(fn ($method) => new ReflectionMethod($model, $method))
-            ->reject(fn (ReflectionMethod $method) => $method->isStatic()
-                || $method->isAbstract()
-                || $method->getDeclaringClass()->getName() !== get_class($model)
+            ->reject(
+                fn (ReflectionMethod $method) => $method->isStatic()
+                    || $method->isAbstract()
+                    || $method->getDeclaringClass()->getName() !== get_class($model)
             )
             ->filter(function (ReflectionMethod $method) {
                 $file = new SplFileObject($method->getFileName());
@@ -480,37 +474,5 @@ class ShowModelCommand extends Command
         return is_dir(app_path('Models'))
             ? $rootNamespace.'Models\\'.$model
             : $rootNamespace.$model;
-    }
-
-    /**
-     * Install the command's dependencies.
-     *
-     * @return void
-     *
-     * @throws \Symfony\Component\Process\Exception\ProcessSignaledException
-     */
-    protected function installDependencies()
-    {
-        $command = collect($this->composer->findComposer())
-            ->push('require doctrine/dbal')
-            ->implode(' ');
-
-        $process = Process::fromShellCommandline($command, null, null, null, null);
-
-        if ('\\' !== DIRECTORY_SEPARATOR && file_exists('/dev/tty') && is_readable('/dev/tty')) {
-            try {
-                $process->setTty(true);
-            } catch (RuntimeException $e) {
-                $this->components->warn($e->getMessage());
-            }
-        }
-
-        try {
-            $process->run(fn ($type, $line) => $this->output->write($line));
-        } catch (ProcessSignaledException $e) {
-            if (extension_loaded('pcntl') && $e->getSignal() !== SIGINT) {
-                throw $e;
-            }
-        }
     }
 }
